usm: Add USM configuration tests
Add comprehensive tests for USM configuration in separate test file,
organized by protocol sections to match the USMConfig structure.

Tests include:
- Global USM configuration (event stream, kernel buffers, concurrent requests)
- HTTP protocol configuration (monitoring, stats, timeouts, replace rules)
- HTTP2 protocol configuration (monitoring, dynamic table cleaner)
- Kafka protocol configuration (monitoring, stats buffered)
- Postgres protocol configuration (monitoring, stats, telemetry)
- Redis protocol configuration (monitoring, track resources, stats)
- TLS configuration by library (Native, Go, NodeJS, Istio)

All tests use direct field access via USMConfig embedding.
